pyproject.toml
[project]
name = "keyremapper"
version = "0.1.0"
description = "Keyboard Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard==0.13.5",
    "pyyaml==6.0.2",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.15.0",
]
